Modify prompt page, to make it up to date with backend
