+++ frontend/components/prompt/chat/chat.module.css
@@ -7,7 +7,7 @@
 }
 
 .userMessage {
-  @apply flex w-full justify-center;
+  @apply flex w-full justify-center pt-16;
 }
 
 .userMessageContent {
